# OpenAPI 3.0.json
openapi: 3.0.3
info:
  title: Multi-tenant Commerce Ingestion API
  version: 1.0.0
  description: >
    CRUD + search APIs for Organizations, Websites, Orders, Fulfillments, and Tracking.
    Supports date-time range filtering and pagination (page/size).
servers:
  - url: https://api.example.com
...
paths:
...
  /orders:
    post:
      tags: [Orders]
      summary: Create (or upsert) order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreateRequest"
      responses:
        "201":
          description: Created/Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      tags: [Orders]
      summary: Search orders (date range + pagination)
      operationId: searchOrders
      parameters:
        - name: orgId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: websiteId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/from"
        - $ref: "#/components/parameters/to"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/OrderStatus"
        - name: financialStatus
          in: query
          schema:
            $ref: "#/components/schemas/FinancialStatus"
        - name: fulfillmentStatus
          in: query
          schema:
            $ref: "#/components/schemas/FulfillmentOverallStatus"
      responses:
        "200":
          description: Paged orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedOrder"
        "400":
          $ref: "#/components/responses/BadRequest"
  /orders/search:
    get:
      tags: [Orders]
      summary: Search order by external order id/number
      operationId: searchOrderByExternal
      parameters:
        - name: orgId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: websiteId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: externalOrderId
          in: query
          required: false
          schema:
            type: string
        - name: externalOrderNumber
          in: query
          required: false
          schema:
            type: string
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
      responses:
        "200":
          description: Paged orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedOrder"
        "400":
          $ref: "#/components/responses/BadRequest"
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order by id
      operationId: getOrderById
      parameters:
        - $ref: "#/components/parameters/orderIdPath"
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Orders]
      summary: Update order (full replace)
      operationId: updateOrder
      parameters:
        - $ref: "#/components/parameters/orderIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Orders]
      summary: Update order (partial)
      operationId: patchOrder
      parameters:
        - $ref: "#/components/parameters/orderIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderPatchRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Orders]
      summary: Delete order
      operationId: deleteOrder
      parameters:
        - $ref: "#/components/parameters/orderIdPath"
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"
  /orders/{orderId}/fulfillments:
    post:
      tags: [Fulfillments]
      summary: Create fulfillment for an order
      operationId: createFulfillment
      parameters:
        - $ref: "#/components/parameters/orderIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FulfillmentCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fulfillment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags: [Fulfillments]
      summary: List/search fulfillments for an order (date range + pagination)
      operationId: listFulfillments
      parameters:
        - $ref: "#/components/parameters/orderIdPath"
        - $ref: "#/components/parameters/from"
        - $ref: "#/components/parameters/to"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/FulfillmentStatus"
        - name: carrier
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paged fulfillments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedFulfillment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /orders/{orderId}/fulfillments/search:
    get:
      tags: [Fulfillments]
      summary: Search fulfillment by external fulfillment id
      operationId: searchFulfillmentsByExternal
      parameters:
        - $ref: "#/components/parameters/orderIdPath"
        - name: externalFulfillmentId
          in: query
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
      responses:
        "200":
          description: Paged fulfillments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedFulfillment"
        "400":
          $ref: "#/components/responses/BadRequest"
  /orders/{orderId}/fulfillments/{fulfillmentId}:
    get:
      tags: [Fulfillments]
      summary: Get fulfillment by id
      operationId: getFulfillmentById
      parameters:
        - $ref: "#/components/parameters/orderIdPath"
        - $ref: "#/components/parameters/fulfillmentIdPath"
      responses:
        "200":
          description: Fulfillment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fulfillment"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Fulfillments]
      summary: Update fulfillment (full replace)
      operationId: updateFulfillment
      parameters:
        - $ref: "#/components/parameters/orderIdPath"
        - $ref: "#/components/parameters/fulfillmentIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FulfillmentUpdateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fulfillment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Fulfillments]
      summary: Update fulfillment (partial)
      operationId: patchFulfillment
      parameters:
        - $ref: "#/components/parameters/orderIdPath"
        - $ref: "#/components/parameters/fulfillmentIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FulfillmentPatchRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fulfillment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Fulfillments]
      summary: Delete fulfillment
      operationId: deleteFulfillment
      parameters:
        - $ref: "#/components/parameters/orderIdPath"
        - $ref: "#/components/parameters/fulfillmentIdPath"
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    from:
      name: from
      in: query
      required: false
      description: Start datetime (inclusive), ISO-8601
      schema:
        type: string
        format: date-time
    to:
      name: to
      in: query
      required: false
      description: End datetime (inclusive), ISO-8601
      schema:
        type: string
        format: date-time
    page:
      name: page
      in: query
      required: false
      description: 0-based page index
      schema:
        type: integer
        minimum: 0
        default: 0
    size:
      name: size
      in: query
      required: false
      description: Batch size / page size
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    sort:
      name: sort
      in: query
      required: false
      description: Sorting, e.g. createdAt,desc or updatedAt,asc
      schema:
        type: string
        default: updatedAt,desc
...
    orderIdPath:
      name: orderId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    fulfillmentIdPath:
      name: fulfillmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
      required: [timestamp, status, error, message, path]

    # ----------------------------
    # Enums
    # ----------------------------
...
    OrderStatus:
      type: string
      enum: [CREATED, CANCELLED, CLOSED]
    FinancialStatus:
      type: string
      enum: [UNKNOWN, PENDING, PAID, PARTIALLY_PAID, REFUNDED, PARTIALLY_REFUNDED, VOIDED]
    FulfillmentOverallStatus:
      type: string
      enum: [UNFULFILLED, PARTIAL, FULFILLED, CANCELLED, UNKNOWN]
    FulfillmentStatus:
      type: string
      enum: [CREATED, SHIPPED, DELIVERED, CANCELLED, FAILED, UNKNOWN]
...
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        websiteId:
          type: string
          format: uuid
        externalOrderId:
          type: string
        externalOrderNumber:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/OrderStatus"
        financialStatus:
          $ref: "#/components/schemas/FinancialStatus"
        fulfillmentStatus:
          $ref: "#/components/schemas/FulfillmentOverallStatus"
        customerEmail:
          type: string
          format: email
          nullable: true
        orderTotal:
          type: number
          format: double
        currency:
          type: string
          nullable: true
          example: USD
        orderCreatedAt:
          type: string
          format: date-time
          nullable: true
        orderUpdatedAt:
          type: string
          format: date-time
          nullable: true
        ingestedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, orgId, websiteId, externalOrderId, status, financialStatus, fulfillmentStatus, orderTotal, ingestedAt, createdAt, updatedAt]
    OrderCreateRequest:
      type: object
      properties:
        orgId:
          type: string
          format: uuid
        websiteId:
          type: string
          format: uuid
        externalOrderId:
          type: string
        externalOrderNumber:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/OrderStatus"
        financialStatus:
          $ref: "#/components/schemas/FinancialStatus"
        fulfillmentStatus:
          $ref: "#/components/schemas/FulfillmentOverallStatus"
        customerEmail:
          type: string
          format: email
          nullable: true
        orderTotal:
          type: number
          format: double
        currency:
          type: string
          nullable: true
        orderCreatedAt:
          type: string
          format: date-time
          nullable: true
        orderUpdatedAt:
          type: string
          format: date-time
          nullable: true
      required: [orgId, websiteId, externalOrderId]
    OrderUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/OrderCreateRequest"
    OrderPatchRequest:
      type: object
      properties:
        externalOrderNumber:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/OrderStatus"
        financialStatus:
          $ref: "#/components/schemas/FinancialStatus"
        fulfillmentStatus:
          $ref: "#/components/schemas/FulfillmentOverallStatus"
        customerEmail:
          type: string
          format: email
          nullable: true
        orderTotal:
          type: number
          format: double
        currency:
          type: string
          nullable: true
        orderCreatedAt:
          type: string
          format: date-time
          nullable: true
        orderUpdatedAt:
          type: string
          format: date-time
          nullable: true
    PagedOrder:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
      required: [data, page, size, totalElements, totalPages, hasNext]

    # ----------------------------
    # Fulfillment
    # ----------------------------
    Fulfillment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        externalFulfillmentId:
          type: string
        status:
          $ref: "#/components/schemas/FulfillmentStatus"
        carrier:
          type: string
          nullable: true
        serviceLevel:
          type: string
          nullable: true
        shippedAt:
          type: string
          format: date-time
          nullable: true
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, orderId, externalFulfillmentId, status, createdAt, updatedAt]
    FulfillmentCreateRequest:
      type: object
      properties:
        externalFulfillmentId:
          type: string
        status:
          $ref: "#/components/schemas/FulfillmentStatus"
        carrier:
          type: string
          nullable: true
        serviceLevel:
          type: string
          nullable: true
        shippedAt:
          type: string
          format: date-time
          nullable: true
        deliveredAt:
          type: string
          format: date-time
          nullable: true
      required: [externalFulfillmentId]
    FulfillmentUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/FulfillmentCreateRequest"
    FulfillmentPatchRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/FulfillmentStatus"
        carrier:
          type: string
          nullable: true
        serviceLevel:
          type: string
          nullable: true
        shippedAt:
          type: string
          format: date-time
          nullable: true
        deliveredAt:
          type: string
          format: date-time
          nullable: true
    PagedFulfillment:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Fulfillment"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
      required: [data, page, size, totalElements, totalPages, hasNext]

    # ----------------------------
    # Tracking
    # ----------------------------
...